+++ modules/apps/portal-search-elasticsearch7/portal-search-elasticsearch7-impl/src/main/java/com/liferay/portal/search/elasticsearch7/internal/facet/DefaultFacetTranslator.java
@@ -14,6 +14,9 @@
 
 package com.liferay.portal.search.elasticsearch7.internal.facet;
 
+import com.liferay.osgi.service.tracker.collections.map.ServiceReferenceMapperFactory;
+import com.liferay.osgi.service.tracker.collections.map.ServiceTrackerMap;
+import com.liferay.osgi.service.tracker.collections.map.ServiceTrackerMapFactory;
 import com.liferay.portal.kernel.search.BooleanClause;
 import com.liferay.portal.kernel.search.Query;
 import com.liferay.portal.kernel.search.facet.Facet;
@@ -36,7 +39,10 @@ import org.elasticsearch.index.query.QueryBuilders;
 import org.elasticsearch.search.aggregations.AggregationBuilder;
 import org.elasticsearch.search.builder.SearchSourceBuilder;
 
+import org.osgi.framework.BundleContext;
+import org.osgi.service.component.annotations.Activate;
 import org.osgi.service.component.annotations.Component;
+import org.osgi.service.component.annotations.Deactivate;
 import org.osgi.service.component.annotations.Reference;
 
 /**
@@ -77,8 +83,7 @@ public class DefaultFacetTranslator implements FacetTranslator {
 			postFilterQueryBuilderOptional.ifPresent(
 				postFilterQueryBuilders::add);
 
-			Optional<AggregationBuilder> optional =
-				_facetProcessor.processFacet(facet);
+			Optional<AggregationBuilder> optional = _processFacet(facet);
 
 			optional.map(
 				aggregationBuilder -> postProcessAggregationBuilder(
@@ -92,6 +97,29 @@ public class DefaultFacetTranslator implements FacetTranslator {
 			searchSourceBuilder.postFilter(
 				_getPostFilter(postFilterQueryBuilders));
 		}
+	}
+
+	@Activate
+	protected void activate(BundleContext bundleContext) {
+		_serviceTrackerMap = ServiceTrackerMapFactory.openSingleValueMap(
+			bundleContext,
+			(Class<FacetProcessor<SearchRequestBuilder>>)
+				(Class<?>)FacetProcessor.class,
+			null,
+			ServiceReferenceMapperFactory.create(
+				bundleContext,
+				(facetProcessor, emitter) -> {
+					String facetClassName = facetProcessor.getFacetClassName();
+
+					if (facetClassName != null) {
+						emitter.emit(facetClassName);
+					}
+				}));
+	}
+
+	@Deactivate
+	protected void deactivate() {
+		_serviceTrackerMap.close();
 	}
 
 	protected AggregationBuilder postProcessAggregationBuilder(
@@ -130,6 +158,19 @@ public class DefaultFacetTranslator implements FacetTranslator {
 		return boolQueryBuilder;
 	}
 
+	private Optional<AggregationBuilder> _processFacet(Facet facet) {
+		Class<?> clazz = facet.getClass();
+
+		FacetProcessor<SearchRequestBuilder> facetProcessor =
+			_serviceTrackerMap.getService(clazz.getName());
+
+		if (facetProcessor == null) {
+			facetProcessor = _defaultFacetProcessor;
+		}
+
+		return facetProcessor.processFacet(facet);
+	}
+
 	private QueryBuilder _translateBooleanClause(
 		BooleanClause<Filter> booleanClause) {
 
@@ -154,10 +195,13 @@ public class DefaultFacetTranslator implements FacetTranslator {
 		return Optional.of(queryBuilder);
 	}
 
-	@Reference(service = CompositeFacetProcessor.class)
-	private FacetProcessor<SearchRequestBuilder> _facetProcessor;
+	@Reference(service = DefaultFacetProcessor.class)
+	private FacetProcessor<SearchRequestBuilder> _defaultFacetProcessor;
 
 	@Reference(target = "(search.engine.impl=Elasticsearch)")
 	private FilterTranslator<QueryBuilder> _filterTranslator;
+
+	private ServiceTrackerMap<String, FacetProcessor<SearchRequestBuilder>>
+		_serviceTrackerMap;
 
 }
